+++ assets/index.html
@@ -8,26 +8,26 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Fira+Code&family=Fira+Sans&display=swap" rel="stylesheet">
 
-  <script src="xterm.js"></script>
-  <script src="xterm-addon-attach.js"></script>
-  <script src="xterm-addon-fit.js"></script>
-  <script src="xterm-addon-web-links.js"></script>
+  <script src="/assets/external/xterm.js"></script>
+  <script src="/assets/external/xterm-addon-attach.js"></script>
+  <script src="/assets/external/xterm-addon-fit.js"></script>
+  <script src="/assets/external/xterm-addon-web-links.js"></script>
 
-  <script src="main.js"></script>
+  <script src="/assets/main.js"></script>
 
-  <link rel="stylesheet" href="xterm.css" />
-  <link rel="stylesheet" href="main.css" />
+  <link rel="stylesheet" href="/assets/external/xterm.css" />
+  <link rel="stylesheet" href="/assets/main.css" />
 
-  <title>Websocket Terminal</title>
+  <title>{{.title}}</title>
 </head>
 
 <body>
-  <h2>Web Terminal</h2>
+  <h2>{{.title}}</h2>
   <div id="terminal">
     <div id="terminal_view"></div>
   </div>
   <script>
-    term = createTerminal();
+    term = createTerminal({{.path}});
     // print something to test output and scroll
     var str = [
       ' ┌────────────────────────────────────────────────────────────────────────────┐',
