+++ assets/main.js
@@ -1,4 +1,4 @@
-function createTerminal() {
+function createTerminal(path) {
   // vscode-snazzy https://github.com/Tyriar/vscode-snazzy
   // copied from xterm.js website
   var baseTheme = {
@@ -43,7 +43,7 @@ function createTerminal() {
   fitAddon.fit();
 
   // create the websocket and connect to the server
-  const ws_uri = "wss://" + window.location.host + "/ws";
+  const ws_uri = "wss://" + window.location.host + path;
   const socket = new WebSocket(ws_uri);
   const attachAddon = new AttachAddon.AttachAddon(socket);
   term.loadAddon(attachAddon);
